Updated Game Project
Scene "Room_0" was created. "ManageMenu" script was created, so that "Player" no longer manages the menu, but the room does. This fix the menu scale and position problem (it was relative to the players scale and position :neutral_face: )
Also, "Player" animation was improved, by remaking the "set_anim()" function!