Updated Game Project
Now, the "Bag" scene no longer controls item usage, it has a "get_item" function.
The "Menu" scene now has "Bag" functions: "get_bag()" and "get_item()". The first return the "bag" variable (which is used to tell if the bag is on) and the second one uses the "Bag" scene function to return an item.
The "ManageMenu" script is now responsible for handling item usage. If the menu's "bag" variable is on, and the "interact_menu" input is pressed, it uses the item, by using the "get_item()" function!