Updated Player Script
Now the "use_item()" function uses the "Item" script functions to make changes to the player variable!
The next change I have to make is in the "Bag" script, where instead of storing lines, it will store variables. This way, I can insert instances of "Item" in it, so that the player's "use_item()" function can work properly...